# docs/swagger.yaml
basePath: /
definitions:
  categories.Categories:
    additionalProperties:
      $ref: '#/definitions/categories.Category'
    type: object
  categories.Category:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  models.IndexStatus:
    properties:
      errors:
        items:
          type: string
        type: array
      index:
        type: string
      is_aggregate:
        type: boolean
      size:
        type: integer
    type: object
  models.LatestResult:
    properties:
      desc:
        type: string
      link:
        type: string
      name:
        type: string
      site:
        type: string
    type: object
  search.Capability:
    properties:
      available:
        type: boolean
      key:
        type: string
      supportedParams:
        items:
          type: string
        type: array
    type: object
  server.statusResponse:
    properties:
      indexes:
        items:
          $ref: '#/definitions/models.IndexStatus'
        type: array
      latest:
        items:
          $ref: '#/definitions/models.LatestResult'
        type: array
    type: object
  torznab.Capabilities:
    properties:
      categories:
        $ref: '#/definitions/categories.Categories'
      searchModes:
        items:
          $ref: '#/definitions/search.Capability'
        type: array
    type: object
  torznab.Info:
    properties:
      category:
        type: string
      description:
        type: string
      id:
        type: string
      language:
        type: string
      link:
        type: string
      title:
        type: string
    type: object
  torznab.ResultFeed:
    properties:
      info:
        $ref: '#/definitions/torznab.Info'
      items:
        items: {}
        type: array
    type: object
  torznab.err:
    properties:
      code:
        type: integer
      description:
        type: string
    type: object
host: localhost:5000
info:
  contact: {}
  description: Torrentd is a torrent RSS feed generator.
  title: Torrentd API
  version: 1.0.0
paths:
  /status:
    get:
      consumes:
      - '*/*'
      description: get status of the server
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.statusResponse'
      summary: Status of the server
      tags:
      - status
  /torznab/{indexes}:
    get:
      consumes:
      - '*/*'
      description: Query indexes in torznab format
      parameters:
      - default: all
        description: Index name(s) to search through
        in: path
        name: indexes
        type: string
      - description: Type of search. Can be caps, search, tvsearch, tv-search, movie,
          movie-search, moviesearch. Defaults to caps, returning the capabilities.
        in: query
        name: t
        type: string
      - description: Search query
        in: query
        name: q
        type: string
      - description: Category
        in: query
        name: cat
        type: string
      - description: The output format to use
        in: query
        name: format
        type: string
      - description: IMDB ID
        in: query
        name: imdbid
        type: string
      - description: TMDB ID
        in: query
        name: tmdbid
        type: string
      - description: TVDB ID
        in: query
        name: rid
        type: string
      - description: Season number
        in: query
        name: season
        type: string
      - description: Episode number
        in: query
        name: ep
        type: string
      - description: Limit the number of results, defaults to 20
        in: query
        name: limit
        type: string
      - description: Offset the results
        in: query
        name: offset
        type: string
      - description: Minimum age of the torrent
        in: query
        name: minage
        type: string
      - description: Maximum age of the torrent
        in: query
        name: maxage
        type: string
      - description: Minimum size of the torrent
        in: query
        name: minsize
        type: string
      - description: Maximum size of the torrent
        in: query
        name: maxsize
        type: string
      - description: API key
        in: query
        name: apikey
        type: string
      produces:
      - text/xml
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/torznab.ResultFeed'
        "404":
          description: 404 page not found
          schema:
            type: type
      summary: Torznab handler
      tags:
      - torznab
  /torznab/caps/{indexes}:
    get:
      consumes:
      - '*/*'
      description: Get the index(es) capabilities in torznab format
      parameters:
      - description: Index name(s) to get the capabilities for
        in: path
        name: indexes
        type: string
      produces:
      - text/xml
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/torznab.Capabilities'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/torznab.err'
      summary: Torznab capabilities
      tags:
      - torznab
schemes:
- http
swagger: "2.0"